requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
bcrypt==5.0.0
certifi==2026.2.25
click==8.3.1
coverage==7.13.5
Faker==40.11.0
iniconfig==2.3.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.1.0
pytest-mock==3.15.1
python-dotenv==1.2.2
rich==14.3.3
sentry-sdk==2.55.0
shellingham==1.5.4
SQLAlchemy==2.0.48
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
